rag-multimodel/requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
pydantic-settings==2.11.0
langchain==1.0.0
langchain-community==0.4
langchain-huggingface==1.0.0
faiss-cpu==1.12.0
ollama==0.6.0
pypdf==4.3.1
langchain-text-splitters==1.0.0
langchain-ollama
pymupdf
numpy
opencv-python-headless
python-multipart
sentence-transformers
